Allowed more forward reasoning
Before: A forward rule must be either simplifying or and elimination
rule in order to get applied. I.e. only elimination rules are allowed
to increase the node count of a derived consequence.

Now: Some rules are regarded as elimination rules even if the don't
eliminate any symbol. This is the case if an elimination rule is added
by partial specialization. The elimination property carries over to this
partial specialization and to the consequence. The 'elim' flag is resetted
as soon as all variables are substituted.

	modified:   ocaml/proof_context.ml