Implemented improved focussed expansion
The improved version expands arguments if the function term doesn't
have an expansion.

However the improved version might do a lot of expansions which are
not useful. E.g.

     (not a => false) => not not a
     ~>   (not a => false) => not a => false

because '=>' does not have an expansion. But this is not a useful
expansion.

Therefore the current checking just includes the code but does not
use it. Only function terms can be in the focus and if the function
term does not have an expansion then the whole term does not have an
expansion.

	modified:   ocaml/feature_table.ml