Decoupled public and private view in class table
The public and private view in the class table are now completely
decoupled. This applies to classes to features and assertions of
classes and to descendants and ancestors.

Each addition/modification made in the private view is accessible
only in the private view, each addition/modification made in the
public view is available in both views.

Builtin classes are inserted only to the private view. The same
applies to builtin functions.

The same decoupling still has to be done for the feature table as
well in order to be able to verify interface files.

	modified:   ocaml/class_table.ml
	modified:   ocaml/class_table.mli
	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/eiffelc.ml
	modified:   ocaml/feature_table.ml
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_context.mli