Some cleanups in the parser and the lexer
The module Support.Parse_info is no longer necessary to process error
messages during parsing and lexing.

The lexer calls 'error_info' directly because it has all position
in case of a lexing error.

The parser just raises 'Parsing.Parse_error' and the handler gets the
position of the last token (the unexpected token) from the lexer.

	modified:   ocaml/eiffelc.ml
	modified:   ocaml/lexer.mll
	modified:   ocaml/parser.mly
	modified:   ocaml/support.ml