Fix: Features do not arrive in ascending sequence
In the interface file the features can be defined in a different
sequence than in the implementation file. The class table stores
in each class the deferred and effective features in a list and in
a term table. But a term table requires that the terms added to the
table arrive in ascending sequence.

The bug has been fixed by introducing a wrapper for the term table which
stores the indices as they arrive in an array appending it to the end
and puts the array index of the appended element (which is guaranteed to
be ascending) into the term table.

	modified:   ocaml/class_table.ml