A fix and check deferred features and assertions
Fix: All assertions which are exported in the interface file but
already present in the implementation file must be put into the
class table (deferred and effective assertions) and inherited to
descendants of the owner class.

All deferred features and assertions must be exported. This is now
checked at the end of the verification of interface files.

	modified:   ocaml/container.ml
	modified:   ocaml/feature_table.ml
	modified:   ocaml/feature_table.mli
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_context.mli
	modified:   ocaml/proof_table.ml
	modified:   ocaml/proof_table.mli