Prepared term level calculation and refactored backward analysis of terms
1. There is still a potential infinite loop in forward and backward reasoning.
Currently a term is considered not more complicated than another if it has
not more nodes than the other. This should avoid that terms can grow
unlimited in the prover.

But due to function expansions it might be the case that e.g. a premise
has fewer nodes than or the same number of nodes as the target in
backward reasoning, but the premise allows more function expansions
than the target.

In order to handle this case properly we introduce for each function a
level. Basic functions have level 0 and function with definitions have
a level one more than the highest level of the used functions in its
definition.

This level calculation has been prepared but not yet activated.

2. Analysis of backward terms have been done in two places. The common part
has been factored out.

	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/feature_table.ml
	modified:   ocaml/feature_table.mli
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_table.ml
	modified:   ocaml/prover.ml
	modified:   ocaml/typer.ml