Term levels integrated into backward reasoning
Now a premise of a backward rule is considered not more complicated
than the target if it doesn't have a higher level (i.e. function level) and if
it doesn't have more nodes.

Basic functions have level zero, functions with a definition have a level
which is one plus the level of the definition term.

	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_context.mli
	modified:   ocaml/prover.ml