Fix: Replace term levels by node count of expanded term
The algorithm to break infinite loops with term levels does not work.
With the term level algorithm an expression has been considered as
not more complicated than another if it has not more nodes and not
a higher term level.

However a term might be more complicated as another if it has one more
expandable function than the other but the other term already contains
a function with the same level. Therefore the term level does not grow
but the expanded term might be more complex than the expansion of the
other.

Therefore we use the node count of the fully expanded term as a complexity
measure.

	modified:   ocaml/proof_context.ml
	modified:   ocaml/prover.ml