Added 'ANY.eq' and 'PREDICATE.()' as basic functions
This required that the feature map in the feature table has to
use a wrapper of the term table (now called 'Term_table2') like
already used in the class table. Reason: It is no longer guaranteed
that the features arrive in ascending sequence. E.g. by the time
of the definition of the function ANY.= in the source file there
are already features with higher index than the builtin index of
the equality function of ANY.

	modified:   ocaml/class_table.ml
	modified:   ocaml/class_table.mli
	modified:   ocaml/feature_table.ml
	new file:   ocaml/term_table2.ml
	new file:   ocaml/term_table2.mli