Fix: Allowed complete forward reasoning with a backward rule
A backward rule might be not simplifying in forward direction. But
it must be always possible to complete the forward reasoning with
a backward rule. For this purpose it has been inserted. The prover
iterates through all premises of a backward rule. Each proved premise
triggers one forward reasoning until the goal is in the context.

Consider the backward rule 'a => b => c' which has been inserted for
the goal 'c'. The prover tries to prove 'a' and in case of success puts
it into the context. The forward close puts then 'b => c' into the context.
Then then prover tries to prove 'b' puts it into the context. The forward
close then puts 'c' into the context and we are ready. If the forward
reasoner blocks for any reason the insertion of 'b => c' then the process
cannot be terminated successfully. Therefore there must not be any blocking
for the intermediate steps of processing a backward rule for a certain
target.

	modified:   ocaml/proof_context.ml