Catch all rules deactivated
Catch all rules are possible as backward rules on any target. The
current implementation of the prover can cause combinatorial
explosion in the presence of more than one catch all rule.

In case of failure it tries the catch all rules in all possible
permutations. Before activating catch all rules again a solution
has to be found to avoid this combinatorial explosion.

	modified:   ocaml/rule_data.ml