Improved evaluation structure for recursion
The type 'Eval.t' now mimics the structure of 'Term.t' with some
added components to describe evaluations.

The type 'Eval.t' allows to implement full function expansions for
for case that the toplevel function does not have a definition and
is not a boolean function (nor 'all' nor 'some').

	modified:   ocaml/proof.ml
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_table.ml
	modified:   ocaml/proof_table.mli