Introduced prover strength
With 'strength = 0' the prover does no nested exploration of alternatives.
I.e. it enters the goal explores the alternatives to prove the goal but
during the exploration of an alternative no more alternatives are explored
to prove the premises.

A prover with zero strength is applied to verify interface files.

	modified:   ocaml/ast.ml
	modified:   ocaml/prover.ml