Disallowed silly backward rules
Rules of the form

    all(a,b) a*b = a ==> b*a = b ==> a = b

are not used as backward rules because a premise has the same structure
as the target i.e. can be unified with the target and the substitution
is nontrivial. For first premise of the above case we get the substitution

    a ~> a*b,   b ~> a

i.e. at the position of the variable 'a' in the target appears a term
which is just blown up. Since the premises have the same structure as
the target the prover shall not try this rule for backward reasoning.

	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/feature_table.ml
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_table.ml
	modified:   ocaml/rule_data.ml
	modified:   ocaml/term.ml