Namespaces for classes implemented
A fully qualified classname has a library part, a module and a
class name

    lib.mdl.CN

Any class whose classname is identical with the module name ignoring
character cases is the main class of a module. For main classes the
modulname is always optional.

The libary name is almost always optional as long as no shadowing
occurs with duplicate module names in different libraries.

	modified:   ocaml/alba.ml
	modified:   ocaml/ast.ml
	modified:   ocaml/class_table.ml
	modified:   ocaml/class_table.mli
	modified:   ocaml/module_table.ml
	modified:   ocaml/module_table.mli
	modified:   ocaml/proof_context.ml
	modified:   ocaml/support.ml