Refactored feature inheritance
Feature inheritance cannot be done completely in the module 'Feature_table'
because in case of a feature redefinition it has to be proved that
the original definition in the ancestor and the definition in the heir
are equivalent. Therefore access to the prover is needed and the module
'Feature_table' cannot have access to the prover.

Therefore the basic control functions for feature inheritance have been
put into a new module 'Inherit' which has access to the prover.

	modified:   ocaml/ast.ml
	modified:   ocaml/class_table.ml
	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/feature_table.ml
	modified:   ocaml/feature_table.mli
	new file:   ocaml/inherit.ml
	new file:   ocaml/inherit.mli
	modified:   ocaml/parser.mly
	modified:   ocaml/support.ml