Fix: Correct handling of formal generics in quantified expressions
If the typer enters a context for a quantified expression with a
formal generic in the argument list then the formal generic is added
to the context.

However at completion time of the quantified expression the formal
generic can be removed only from the context and not from the term
builders. This is because the builders can contain global type variables
which might refer to the formal generics (which in the term builder
are considered as type constants).

The solution is to leave all formal generics once entered into the
builder in the builder as garbage like all global type variables.

	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/signature.ml
	modified:   ocaml/term_builder.ml
	modified:   ocaml/term_builder.mli
	modified:   ocaml/tvars.ml
	modified:   ocaml/tvars.mli
	modified:   ocaml/typer.ml