Added code to detect transitivity rules
Basic idea: use transitivity rules only in forward direction even
if they are not strictly simplifying.

Unfortunately the runtime does not improve if we activate this feature.
Therefore the code is just in and not activated to guide the prover.

The inefficiency in some cases has to be analyzed more in depth to
find a good solution.

	modified:   ocaml/rule_data.ml
	modified:   ocaml/rule_data.mli
	modified:   ocaml/term.ml