Added code to detect base features for compliance check with '='
Base features of a class are features which do not have a definition.

If the class inherits from ANY then all these features have to be
consistent with equality so that the leibniz rule is valid.

The detection of base features has been implemented but the
consistency check with '=' is not yet implemented.

	modified:   ocaml/ast.ml
	modified:   ocaml/class_table.ml
	modified:   ocaml/class_table.mli
	modified:   ocaml/feature_table.ml
	modified:   ocaml/feature_table.mli
	modified:   ocaml/inherit.ml
	modified:   ocaml/inherit.mli