Consistent treatment of @DUMMY, PREDICATE and FUNCTION
A dummy type is entered in case that any result or variable which has
been typed with a type variable has to be callable. In case that the
type variable has a concept the dummy type is accepted. But when the
dummy type is upgraded either to a predicate or a function then the
satisfaction of the concept of the type variable is checked. In case
that the concept is not satisfied the term builder does not accept
the expression.

	modified:   ocaml/class_table.ml
	modified:   ocaml/term_builder.ml
	modified:   ocaml/term_builder.mli
	modified:   ocaml/typer.ml