Enriched term structure
Added a predicate flag to lambda terms and applications.

a) lambda terms: The flag indicates if the lambda term represents
a predicate or a function. Without this flag this distinction is no
longer visible in the term.

b) applications: The flag indicates if an applied untyped variable
represents a predicate or a function. Without this flag this distinction
is not possible having the term and its context only.

Caveat: The flag in the applications is not yet filled nor used.

	modified:   ocaml/class_table.ml
	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/feature_table.ml
	modified:   ocaml/feature_table.mli
	modified:   ocaml/proof.ml
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_table.ml
	modified:   ocaml/signature.ml
	modified:   ocaml/support.ml
	modified:   ocaml/term.ml
	modified:   ocaml/term_builder.ml
	modified:   ocaml/term_builder.mli
	modified:   ocaml/term_table.ml
	modified:   ocaml/term_table.mli
	modified:   ocaml/tvars.ml
	modified:   ocaml/typer.ml