Tried to use the type checker during term specialization
The code has been left in but it is not activated, because some rework
of the type system might be necessary.

Goal: Provide for each term a context where all variables of the term
have a defined type.

Steps:

1. The context has to be made immutable, because inner lambda terms
   need a new context where the bound variables get a type

2. The lambda terms have to carry type information.

Having this it is possible to check for each subterm if it is well
typed which in the current design is not possible.

Furthermore it would be possible to define a type for each substitution
and in doing the substitution the type validity could be checked and
potential specializations could be triggered.

	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_table.ml
	modified:   ocaml/rule_data.ml
	modified:   ocaml/term_builder.ml
	modified:   ocaml/term_builder.mli
	modified:   ocaml/typer.ml