Concept of an anchor class introduced
A feature/assertion has an anchor class if it has exactly one formal
generic appearing in the signature on the toplevel. The anchor class
is the principal class of the concept of this formal generic.

Features/assertions are inherited via the anchor class as a parent class.

E.g. with G:ANY the feature

    singleton (a:G): G?

is anchored to ANY, i.e. all descendants of ANY get a variant of this feature.

The assertion

    all(x:G, p:G?) p(x) ==> {x} <= p

is anchored to ANY as well and therefore inherited to all descendants of
ANY.

	modified:   ocaml/ast.ml
	modified:   ocaml/class_table.ml
	modified:   ocaml/class_table.mli
	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/feature_table.ml
	modified:   ocaml/inherit.ml
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_context.mli
	modified:   ocaml/prover.ml
	modified:   ocaml/prover2.ml
	modified:   ocaml/tvars.ml