Term simplification before term evaluation (expansion)
If an assertion/goal can be simplified because it has some subterm
which is equal to a simpler term (simpler in terms of node count) then
the simplification is done and not any expansion of definitions
(neither top level nor any other) is done in this step.

Function expansion is done only if the term cannot be simplified any
more.

	modified:   ocaml/proof_context.ml