Allowed arbitrary expression for a function in dot notation
We allow object oriented syntax in the form

    a.closed(p)

as a synonym for

   closed(a,p)

in mathematical notation.

In order to allow arbitrary expressions at the position where the
function ('closed' in the example) occurs we augmented the parser
to accept

    a.(expr)(p)

where 'expr' can be any expression which represents a two argument
function with the proper type.

	modified:   ocaml/parser.mly