Added signatures and some preparations for precondition checking
1. Added signatures

The term builder now builds the terms together with their signatures.
This is a first step to integrate the types of the arguments into
lambda expressions. The types have to be inserted into lambda expressions
because if not they cannot be reconstructed in some cases. E.g. the
function literal

     ((x:SOMETYPE) -> exp)  -- exp does not contain 'x'

will be converted into a lambda expression. If the lambda expression does
not contain the type 'SOMETYPE' this type is definitely lost.

2. Precondition checking

A basic procedure has been added. But further implementation is stuck
until type information is available in lambda expressions.

	modified:   ocaml/ast.ml
	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/feature.ml
	modified:   ocaml/feature.mli
	modified:   ocaml/feature_table.ml
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_table.ml
	modified:   ocaml/proof_table.mli
	modified:   ocaml/term_builder.ml
	modified:   ocaml/term_builder.mli
	modified:   ocaml/typer.ml