Added functions to tupelize lambda expressions
All lambda expressions must have one argument. If there is more than
one argument, all arguments have to be bundled into a tuple.

The function 'tupelize' bundles all arguments into a tuple and the
function 'untupelize' unpacks the tuple to individual arguments.

The functions are implemented and tested but not yet used.

	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/feature_table.ml
	modified:   ocaml/feature_table.mli
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_context.mli
	modified:   ocaml/proof_table.ml
	modified:   ocaml/proof_table.mli
	modified:   ocaml/term.ml