Term redesign (2)
Removed the operators 'all' and 'some' from the feature table.

We added a flag 'is_all' to the term constructor 'QLam' to indicate
if the quantified expression is universally or existentially quantified.
Therefore the operators 'all' and 'some' which mapped a predicate to
a boolean are no longer necessary. All information abount quantified
expressions is completeley contained in 'QLam(nargs,names,inner_term,is_all)'.

	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/feature_table.ml
	modified:   ocaml/feature_table.mli
	modified:   ocaml/proof.ml
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_table.ml
	modified:   ocaml/proof_table.mli
	modified:   ocaml/rule_data.ml
	modified:   ocaml/term.ml
	modified:   ocaml/term_algo.ml
	modified:   ocaml/term_builder.ml
	modified:   ocaml/term_builder.mli
	modified:   ocaml/term_table.ml
	modified:   ocaml/typer.ml