Term redesign (6)
Lambda terms with one argument only.

Function and predicate application have one arguments only. If there is
more than one argument the sequence of arguments is packed into a tuple.

In order to access the tuple elements, tuple accessors have to be added
to the inner term of a lambda term.

	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/feature_table.ml
	modified:   ocaml/feature_table.mli
	modified:   ocaml/inherit.ml
	modified:   ocaml/proof.ml
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_table.ml
	modified:   ocaml/proof_table.mli
	modified:   ocaml/prover2.ml
	modified:   ocaml/term.ml
	modified:   ocaml/term_algo.ml
	modified:   ocaml/term_builder.ml
	modified:   ocaml/term_builder.mli
	modified:   ocaml/term_table.ml
	modified:   ocaml/typer.ml