Used the concept of an anchor class to inherit assertions
In any proof it is allowed to use more variables than needed for a specific
assertion. This is used heavily in interface files to get a readable documentation
of a module. The compiler throws out all unused variables, type checks the
term and calculates the anchor class of the assertion if there is one.

In order to find an anchor class the compiler has to find an anchor formal
generic first. An anchor formal generic is a formal generic which is used
directly to declare a variable. The anchor formal generic exists if there is
only one anchor formal generic and this is the only formal generic used in
the whole assertion.

The anchor class is the principal class of the anchor formal generic.

However the concept is still not general enough to cover of instantiation
cases of assertions, especially of more than one formal generic is involved.

In the long run we have to add type information to the terms to resolve this
problem.

	modified:   ocaml/ast.ml
	modified:   ocaml/class_table.ml
	modified:   ocaml/context.ml
	modified:   ocaml/feature_table.ml
	modified:   ocaml/inherit.ml
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_context.mli
	modified:   ocaml/prover2.ml
	modified:   ocaml/rule_data.ml
	modified:   ocaml/rule_data.mli
	modified:   ocaml/signature.ml
	modified:   ocaml/term.ml
	modified:   ocaml/term_builder.ml
	modified:   ocaml/term_builder.mli