Added result variable ('Result') to contexts
This is a preparation for functions defined by properties.

For functions defined by properties we need a variable in the function context
which represents the result term. Therefore functions can have 'nargs + 1'
variables, the last one representing the result term.

In a context we now clearly distiguish between variables and arguments. For a
specific context the number of variables is either equal to the number of
arguments or one more than the number of arguments depending on the optional
creation of a result variable.

	modified:   ocaml/ast.ml
	modified:   ocaml/class_table.ml
	modified:   ocaml/class_table.mli
	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/feature.ml
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_context.mli
	modified:   ocaml/proof_table.ml
	modified:   ocaml/proof_table.mli
	modified:   ocaml/rule_data.ml
	modified:   ocaml/term_builder.ml
	modified:   ocaml/typer.ml