Fix: Inconsistent type in type checker
The type checker upgrades a callable signature of the form

    (A):RT

to A? or A->RT depending on the predicate flag of the application.

Since A or RT might be type variables without concept the types
A? and A->RT might be inconsistent. Therefore both, the argument
and the result type, are unified with new global type variables with
concept ANY and before upgrading the required signature to a predicate
or function, A and RT will be substituted according to the substitutions
in the term builder.

	modified:   ocaml/term_builder.ml