Catch all rules extended to predicates and functions
The double negation rule 'all(a) not not a  ==>  a' is a catch all rule
in backward direction. Therefore it is not used as a valid backward rule.

The rule 'all(a,b) a ==> b ==> a and b' is a catch all rule for partial
specialization in forward direction. Therefore it is not use as a valid
forward rule.

Now rules with target or first premise 'p(a)' where 'p' and 'a' are variables
are not used a valid rules either. Reason: There are simply too many matches.

Note: Induction rules of the form

    all(p:NATURAL?, x:NATURAL)
        require  p(0)
                 all(n) p(n) ==> p(n.successor)
        ensure   p(x)
        end

are no longer valid backward rules. In order to trigger an induction proof
at least the induction start has to be entered.

	modified:   ocaml/rule_data.ml