Added constructors and rules of inductive types
An inductive type has the form

    case class LIST[G] create
        nil
        (^) (head:G, tail:LIST[G])
    end

For this type of classes the constructors are added to the features
and the following rules are added to the assertions:

    -- induction
    all(p,s) p(nil)  ==>  (all(s,x) p(s) ==> p(x^s))  ==>  p(s)

    -- inversion
    all(x,s)  nil = x^s  ==>  false
    all(x,s)  x^s = nil  ==>  false

    -- injection
    all(x,y,s,t) x^s = y^t  ==> x = y
    all(x,y,s,t) x^s = y^t  ==> s = t

Computation with (i.e. evaluation of) expression with inductive types
are not yet implemented.

Furthermore destructors (e.g. access to 'head' and 'tail' of a list) are
not yet implemented.

	modified:   ocaml/ast.ml
	modified:   ocaml/class_table.ml
	modified:   ocaml/class_table.mli
	modified:   ocaml/context.ml
	modified:   ocaml/feature_table.ml
	modified:   ocaml/feature_table.mli
	modified:   ocaml/parser.mly
	modified:   ocaml/proof_context.mli