Function search only within visible modules
This restricted search is important for using of interfaces. An interface
can be used in an environment with more modules available as in the
environment where the module has been verified. Therefore more functions
are available which can create ambiguities. New ambiguities are avoided
by searching for functions only whithin the modules which are visible to
the module.

	modified:   ocaml/alba.ml
	modified:   ocaml/feature_table.ml
	modified:   ocaml/module_table.ml