Improved dependency tracking of modules
Suppose the interface file of a module has been modified. Then the
Albatross compiler requires recompilation of the module and of all
modules which depend on it. However if 'alba compile <module>' is
issued only the module is recompiled. Further status request showed
that everything was uptodate and the recompilation need of the
dependent modules had been forgotten.

This has been resolved with this change. The dependent modules will
be regarded as affected, because the timestamp of the dali-file of
the dependent modules is older than the interface file of the changed
module.

	modified:   ocaml/alba.ml