Inhibit expansion of functions resulting in flow control
During full expansion (i.e. not for toplevel expansion) a function
definition is not unfolded if it results in a flow control expression
which cannot be further evaluated.

Maybe this restriction has to be applied in the future to recursive
functions only.

Reason: It does not make sense to expand a recursive function unless
        the correct branch can be chosen from the structure of the
	arguments.

	modified:   ocaml/proof_context.ml