Fix: Blocked rules
A rule which is not simplifying in forward direction is not a forward
rule. But if the target does not contain all variables it is not a
backward rule either. The fix resolves the problem that it treats such
a schematic rule as a forward rule, specializes the rule but then
treat the specialized rule as a backward rule.

	modified:   ocaml/proof_context.ml
	modified:   ocaml/rule_data.ml