Revised implication chain sequence of the induction law for inductive sets
-       all(a,q) p(a) ==> ind0 ==> ... ==> indn ==> q(a)
+       all(q,a) ind0 ==> ... ==> indn ==> p(a) ==> q(a)

Consequence: The induction law is not specialized with the presence
of an element in the inductive set (i.e. 'p(a)' valid). Therefore
the presence of a goal of the form 'q(a)' alone cannot start backward
reasoning with the induction law. This reduces a lot of useless
backward reasoning of the prover.

	modified:   ocaml/term.ml