Performance improvement1: induction proofs with inductive sets
The induction law of an inductive set has the following form

    all(q,a) i0 ==> i1 ==> ... ==> p(a) ==> q(a)

where 'p' is the inductive set, 'i0,i1,...' are the induction rules
(one for each case) 'a' is an element of the inductive set and 'q'
is another set which is a superset of 'p' if all induction rules
are satisfied.

The improved algorithm proves 'i0,i1,...,p(a)' without entering the
assertions into the search tables. Via the modus ponens rule the
result 'q(a)' can be proved directly.

	modified:   ocaml/ast.ml
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_context.mli