Performance improvement3: Inductively defined sets not entered directly
Inductively defined sets are rather lengthy expressions. Therefore
inductively defined sets are no longer added directly into the proof
and search tables. Only representants of the sets are entered.

Example:

    even: NATURAL? = {(p): p(0), all(n) p(n) ==> p(n+1)}

An expression of the form

    n in {(p): p(0), all(n) p(n) ==> p(n+1)}

won't be entered into the tables. Instead of this expression the
expression

    n in even

is used. This speeds up the prover significantly.

	modified:   ocaml/ast.ml
	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/feature.ml
	modified:   ocaml/feature.mli
	modified:   ocaml/feature_table.ml
	modified:   ocaml/feature_table.mli
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_context.mli
	modified:   ocaml/proof_table.ml
	modified:   ocaml/proof_table.mli
	modified:   ocaml/term.ml