Variable definitions (1)
Each assertion of the form

    v = exp
or
    exp = v

where 'v' is a variable which does not yet have a definiton and 'v' does
not occur in 'exp' and no variable in 'exp' has a definition is treated
as a definition of the variable 'v'.

Then all assumptions containing the variable 'v' are scanned are the terms
'assumption[v:=exp]' are added to the context and the substituted terms
are added as used terms for all variables in 'exp' (remember that none
of the variables in 'exp' has a definition).

	modified:   library/alba.base/list.al
	modified:   library/alba.base/predicate.al
	modified:   ocaml/feature_table.ml
	modified:   ocaml/feature_table.mli
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_context.mli
	modified:   ocaml/proof_table.ml
	modified:   ocaml/proof_table.mli
	modified:   ocaml/source_prover.ml