Improved module 'term_table' (better performance)
Adding assertions to the tables consumes a large part of the runtime.

Therefore the data structure in the term tables have been improved.
It is no longer necessary to store empty substitutions in the tables.

In unifying function calls the unifier unifies the first argument and
sends down the result of the unification of the first argument to
the unification of the second and soon. I.e. the different branches
are no longer unified separately with the merge afterwards,
now the merge happens in each path by using the previous results.

The performance improvement is 10% approximately.

	modified:   ocaml/term_table.ml