Store the proved assertions only in one table
We store all universally quantified assertions as schematic terms.
Therefore in trying to unify with a term which has argument variables
special care must be taken on how to match the argument variables of
terms in the search table.

The argument variable of terms in the search table can be treated as
local environment variables or bound variables. The latter case applies
if the term unified with is a universally quantified term. Then then
bound variables of the terms in the table are treated as bound variables
if the number of argument variables coincides with the number of quantified
variables.

    benchmark: choice_function: old: 0.44   new: 0.42
               order_relation:  old: 1.34   new: 1.21
               partial_order:   old: 0.72   new: 0.69
               function_logic:  old: 1.81   new: 1.58

               alba.base:       old: 8.77   new: 8.13

	modified:   ocaml/proof_context.ml
	modified:   ocaml/term_table.ml
	modified:   ocaml/term_table.mli