Catchall rules updated
A rule is a forward catchall rule if it is an implication and the
premise is a pure variable of the rule.

A rule ia a backward catchall rule if it is an implication and the
target is a pure variable of the rule or a pure set/relation
expression where both the set/relation and the argument(s) are pure
variables of the rule.

	modified:   ocaml/rule_data.ml